Extracted getConfigType to parser/utils.js
