+++ packages/react-native-codegen/src/parsers/flow/index.js
@@ -15,52 +15,32 @@ import type {SchemaType} from '../../CodegenSchema.js';
 // $FlowFixMe[untyped-import] there's no flowtype flow-parser
 const flowParser = require('flow-parser');
 const fs = require('fs');
-const {visit, buildSchemaFromConfigType} = require('../utils');
+const {buildSchemaFromConfigType, getConfigType} = require('../utils.js');
 const {buildComponentSchema} = require('./components');
 const {wrapComponentSchema} = require('./components/schema');
 const {buildModuleSchema} = require('./modules');
 const {isModuleRegistryCall} = require('./utils');
 
-function getConfigType(
-  // TODO(T71778680): Flow-type this node.
-  ast: $FlowFixMe,
-): 'module' | 'component' | 'none' {
-  let isComponent = false;
-  let isModule = false;
-
-  visit(ast, {
-    CallExpression(node) {
+function Visitor(infoMap: {isComponent: boolean, isModule: boolean}) {
+  return {
+    CallExpression(node: $FlowFixMe) {
       if (
         node.callee.type === 'Identifier' &&
         node.callee.name === 'codegenNativeComponent'
       ) {
-        isComponent = true;
+        infoMap.isComponent = true;
       }
 
       if (isModuleRegistryCall(node)) {
-        isModule = true;
+        infoMap.isModule = true;
       }
     },
-    InterfaceExtends(node) {
+    InterfaceExtends(node: $FlowFixMe) {
       if (node.id.name === 'TurboModule') {
-        isModule = true;
+        infoMap.isModule = true;
       }
     },
-  });
-
-  if (isModule && isComponent) {
-    throw new Error(
-      'Found type extending "TurboModule" and exported "codegenNativeComponent" declaration in one file. Split them into separated files.',
-    );
-  }
-
-  if (isModule) {
-    return 'module';
-  } else if (isComponent) {
-    return 'component';
-  } else {
-    return 'none';
-  }
+  };
 }
 
 function buildSchema(contents: string, filename: ?string): SchemaType {
@@ -73,7 +53,7 @@ function buildSchema(contents: string, filename: ?string): SchemaType {
   }
 
   const ast = flowParser.parse(contents, {enums: true});
-  const configType = getConfigType(ast);
+  const configType = getConfigType(ast, Visitor);
 
   return buildSchemaFromConfigType(
     configType,
