+++ packages/react-native-codegen/src/parsers/typescript/index.js
@@ -14,59 +14,39 @@ import type {SchemaType} from '../../CodegenSchema.js';
 
 const babelParser = require('@babel/parser');
 const fs = require('fs');
-const {visit, buildSchemaFromConfigType} = require('../utils');
+const {buildSchemaFromConfigType, getConfigType} = require('../utils.js');
 const {buildComponentSchema} = require('./components');
 const {wrapComponentSchema} = require('./components/schema');
 const {buildModuleSchema} = require('./modules');
 
 const {isModuleRegistryCall} = require('./utils');
 
-function getConfigType(
-  // TODO(T108222691): Use flow-types for @babel/parser
-  ast: $FlowFixMe,
-): 'module' | 'component' | 'none' {
-  let isComponent = false;
-  let isModule = false;
-
-  visit(ast, {
-    CallExpression(node) {
+function Visitor(infoMap: {isComponent: boolean, isModule: boolean}) {
+  return {
+    CallExpression(node: $FlowFixMe) {
       if (
         node.callee.type === 'Identifier' &&
         node.callee.name === 'codegenNativeComponent'
       ) {
-        isComponent = true;
+        infoMap.isComponent = true;
       }
 
       if (isModuleRegistryCall(node)) {
-        isModule = true;
+        infoMap.isModule = true;
       }
     },
 
-    TSInterfaceDeclaration(node) {
+    TSInterfaceDeclaration(node: $FlowFixMe) {
       if (
         Array.isArray(node.extends) &&
         node.extends.some(
           extension => extension.expression.name === 'TurboModule',
         )
       ) {
-        isModule = true;
+        infoMap.isModule = true;
       }
     },
-  });
-
-  if (isModule && isComponent) {
-    throw new Error(
-      'Found type extending "TurboModule" and exported "codegenNativeComponent" declaration in one file. Split them into separated files.',
-    );
-  }
-
-  if (isModule) {
-    return 'module';
-  } else if (isComponent) {
-    return 'component';
-  } else {
-    return 'none';
-  }
+  };
 }
 
 function buildSchema(contents: string, filename: ?string): SchemaType {
@@ -83,7 +63,7 @@ function buildSchema(contents: string, filename: ?string): SchemaType {
     plugins: ['typescript'],
   }).program;
 
-  const configType = getConfigType(ast);
+  const configType = getConfigType(ast, Visitor);
 
   return buildSchemaFromConfigType(
     configType,
